This commit contains the work-in-progress refactoring of the monolithic MainViewModel into smaller, more focused ViewModels (AuthViewModel, CaseViewModel, EvidenceViewModel, OcrViewModel).
The code is currently in a non-compilable state due to the extensive refactoring. This commit serves as a snapshot of the progress.

The following has been done:
- Created new ViewModels and repositories.
- Updated UI screens to use the new ViewModels.
- Added unit tests for the new components.

The next step is to fix the compilation errors.

+++ app/src/main/java/com/hereliesaz/lexorcist/model/Evidence.kt
@@ -19,8 +19,8 @@ import java.util.Date
 @Entity(tableName = "evidence")
 data class Evidence(
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
-    var caseId: Long = 0, // Added this line
+    val id: Int = 0,
+    var caseId: Int = 0, // Added this line
     val amount: Double?,
     val timestamp: Date,
     val sourceDocument: String,
